refactor: remove unused layout component and update network handling in various functions
