+++ src/common/providers/TokenProvider.tsx
@@ -29,23 +29,25 @@ interface TokenProviderProps {
   children: ReactNode;
   contractAddress?: Address;
   defaultTokenData?: MasterToken;
+  network: string;
 }
 
 export const TokenProvider: React.FC<TokenProviderProps> = ({
   children,
   contractAddress,
   defaultTokenData,
+  network,
 }) => {
   const [tokenData, setTokenData] = useState<MasterToken | null>(
     defaultTokenData || null,
   );
   const [isLoading, setIsLoading] = useState(false);
 
-  const fetchTokenInfo = async (address: string) => {
+  const fetchTokenInfo = async (address: string, network?: string) => {
     setIsLoading(true);
     try {
       console.log("Fetching token data...", address);
-      const tokenResponse = await fetchTokenData(address, null);
+      const tokenResponse = await fetchTokenData(address, null, String(network));
       const clankerResponse = await fetch(
         `/api/clanker/ca?address=${address}`,
       ).then((res) => res.json());
@@ -66,7 +68,7 @@ export const TokenProvider: React.FC<TokenProviderProps> = ({
   // Loads if defaultTokenData is not provided
   useEffect(() => {
     if (!defaultTokenData) {
-      fetchTokenInfo(contractAddress as Address);
+      fetchTokenInfo(contractAddress as Address, network);
     }
   }, []);
