+++ src/constants/etherscanChainIds.ts
@@ -1,5 +1,6 @@
-export const enum EtherScanChains {
+export enum EtherScanChains {
   // from https://docs.etherscan.io/etherscan-v2/getting-started/supported-chains
   "mainnet" = 1,
   "base" = 8453,
+  "polygon" = 137,
 }
