--- src/pages/t/base/[contractAddress]/layout.tsx
@@ -1,7 +1,0 @@
-import React from "react";
-
-function Layout() {
-  return <div>Layout</div>;
-}
-
-export default Layout;
